# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Allizo Service API",
        "description": "API documentation for Allizo Service platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://allizo.example.com",
            "description": "Production Server"
        },
        {
            "url": "http://alizo.test/",
            "description": "Production API Server"
        }
    ],
    "paths": {
        "/api/allizo/booking": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Get all bookings",
                "operationId": "49081b7e165d88dee8835619d2902a91",
                "responses": {
                    "200": {
                        "description": "Bookings retrieved successfully"
                    },
                    "404": {
                        "description": "No bookings found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Create a booking",
                "description": "Store a newly created resource in storage.\n /**",
                "operationId": "555c983be4a17dd6051bceeccadaa79d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully"
                    },
                    "500": {
                        "description": "Booking creation failed"
                    }
                }
            }
        },
        "/api/allizo/booking/{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Get a booking by ID",
                "description": "Display the specified resource.\n /**",
                "operationId": "9ea9fc83d16cbed9e743fcfa82aa1838",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking retrieved successfully"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Booking"
                ],
                "summary": "Update a booking",
                "description": "Update the specified resource in storage.\n/**",
                "operationId": "c6a9e75cd2df4d30d340e12010d2ab63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated successfully"
                    },
                    "500": {
                        "description": "Booking update failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Booking"
                ],
                "summary": "Delete a booking",
                "description": "Remove the specified resource from storage.\n/**",
                "operationId": "c815ebf9ad2f19d2d4978ff68f7a1443",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking deleted successfully"
                    },
                    "500": {
                        "description": "Booking deletion failed"
                    }
                }
            }
        },
        "/api/allizo/booking/{bookingId}/service/{serviceId}": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Add a service to booking",
                "operationId": "5120debf8d64823a1b4965dede25b9e1",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "serviceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_date",
                                    "time_slot",
                                    "status"
                                ],
                                "properties": {
                                    "booking_data": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "time_slot": {
                                        "type": "string",
                                        "format": "H:i:s"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service added successfully"
                    },
                    "400": {
                        "description": "Service already exists"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Booking"
                ],
                "summary": "Remove a service from a booking",
                "operationId": "4214fd5302da06b9fd0deb9778714a7f",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "serviceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service removed from booking successfully"
                    },
                    "404": {
                        "description": "Service not found in this booking"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentication using Laravel Sanctum",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}